Some cleanup and renaming
